Changed evaluation of 'engine' to sustain sprintf format
If 'engine' contains a sprintf formatter markes (%{}) then the check_connection is avoided
and the engine is selected on an eventby event basis.

This changed the way the batch update is done, now instead of a document list,
the documents are partitioned by engine (once resolved by event's sprintf) and then
indexed by each parition.

Fixes #22