Add Workplace Search output
This is mostly copied from the app search output with some changes for
the API and client differences.

I tried to switch to the [shared client][0] but that resulted in a
dependency conflict on `elasticsearch-transport`.

The integration spec is using internal Workplace Search APIs because we
don't currently expose enough to make it work with public APIs (can't
get access token or content source ID). Hopefully this will change in
the near future.

[0]: https://github.com/elastic/enterprise-search-ruby