Rename AppSearch to Enterprise Search
App Search and Workplace Search are shipped together in the Enterprise
Search package. This renames that shared things to use "enterprise" and
the app search things from "appsearch" to "app search" to be consistent.

I think we may want to rename the entire plugin to
`elastic_enterprise_search`, but I'm not sure what that would take.